feat(game): show Open game world map FAB on every page (#1414)
Moves the bottom-right "🗺️ Open game world map" launcher from
app/game/page.tsx into components/AppShell.tsx so it appears on every
page in the app, not just /game.

The FAB sits at z-40; the /game/world page itself uses a fixed inset-0
z-50 overlay, which naturally covers the FAB without needing a path
check. Other minimal-chrome pages (login, signup) currently render
through AppShell too, so the launcher rides along automatically.
